+++ app/src/main/java/woowacourse/shopping/feature/cart/CartActivity.kt
@@ -4,8 +4,8 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import woowacourse.shopping.feature.cart.stateholder.CartStateHolder
-import woowacourse.shopping.feature.cart.ui.theme.AndroidshoppingcartTheme
+import woowacourse.shopping.feature.cart.component.CartScreen
+import woowacourse.shopping.feature.productlist.ui.theme.AndroidshoppingTheme
 
 class CartActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -14,7 +14,7 @@ class CartActivity : ComponentActivity() {
 
         setContent {
             val stateHolder = CartStateHolder(emptyList())
-            AndroidshoppingcartTheme {
+            AndroidshoppingTheme {
                 CartScreen(
                     cartContents = stateHolder.cartContents,
                     onCloseClick = { finish() },
