+++ app/src/main/java/woowacourse/shopping/feature/cart/component/CartAppBar.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.feature.cart
+package woowacourse.shopping.feature.cart.component
 
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
